Back Up Slow Down
For the third time I am trying to rewrite this code. I think I'm getting
there the last major problem I found was that the motor array was
positioned wrong on the last attempt. Also here is the autonBase
function which is designed to direct the robot to move a specific
distance. The next ultility that is going to be added is hopefully being
able to add an angle to autonbase.