Almost There
I finally got the encoders working right. THe issue was two pronged. First the encoders were mounted backwards on the robot so that one encoder was reading backwards. There were also logic issues in the while loop controlling the speed of the motors inside the autonBase() function. 

Right now the issue is that the motors are moving the wrong direction during user control and the only way to get it to driver forward during autonBase() is through a hacky system of reversing some of the motors.