Working Kindof
So it not appears to move in a mostly straight line. (Probably due to the multiplication constant on the proportional controller.) It is purely proportional with no integration or derivative. It moves about 20 cm farther than it should at a target distance of 30cm. So it overshoots by roughly 90%. That is easy to fix all I have to do is fiddle around with the conversion constant. 

In the next update I will have to add turning and probably some sort of control for the arm.