WHYYY.
So we now have code that makes the robot move in a straight line. The real robot seems to be wired different than the robot that I have been testing with. Which makes it incredibly inconvenient. The gyro logic is working on paper as long as the gyro increases its value as it turns left. On code it doesn't work I'm guessing that I need to initialize the gyro, but I don't have the time to do that right now.