Formating Comments
Some of the comments were cleaned up so that they would work with doxygen. I also added the doxygen raw html so that it was a little more transferable between computers. The next step is to add either a page on the robotics website or create a .io of my own. Also on the todo list is to add a more advanced auton (or create one that works at all). As it stands right now the IME registry in this code base is not complete. I am also still running into IME issues.