Gyro is finally working
The title pretty much sums it all up. Still haven't got to the issues outlined in the last update. Still working through all of this. This may very well be the last week I ever work with this code base. Four years of work coming to an end. Wow it doesn't even work very well.

Good news is that I am documenting the PI tuning process for the gyro for teaching the younger students. Thats on the mac laptop main.c file.