Final Update
This is the final code as it ran during the 2016-2017 season for team. No further updates to code will be made to this repository. I'm not sure if I meet all of the goals that I had when I started the season, but I did manage to create my own PID control system so in the end I'm going to call it even.